Updated exchange value  to support decimals

+++ src/Schemas/World/ExchangeEntry.ts
@@ -2,5 +2,5 @@ import { Schema, type } from "@colyseus/schema"
 
 export class ExchangeEntry extends Schema {
     @type ("int32") crypto_id : number = 0;
-    @type ("int32") usd_value: number = 1;
+    @type ("float32") usd_value: number = 1;
 }
